Streamline Echo Broadcast Actor
A single actor handles all Acks for all messages. The actor is started
once and handler is cloned for any client who wants to run echo
broadcast.

The alternative is to demux on message id, and we need a mutex
there. So for now, use an actor to serialized all Acks receives.

The only alternative is to use a demultiplexer on message id. We'll
optimise this if need be.